Adding my CS 235 Coursework.
CS 235 - Object Oriented Programming - Coursework and Projects. This may
be used freely by anyone provided it is not being used to cheat on your
homework or anything related. I do not condone this, nor am I
responsible for anything done with any of this code.

On a brighter note, these programs are written in Java.